+++ doc/changes.rst
@@ -5,6 +5,8 @@ desitarget Change Log
 4.7.2 (unreleased)
 ------------------
 
+* Revert ToO changes between desitarget/4.5.0 and 4.6.0 [`PR #880`_].
+    * This will allow a single tag for both ToO and MTL updates.
 * Catch reprocessing bug where 1B tile list can change size [`PR #879`_].
     * The input tile list can be bigger than the output tile list.
     * This is because not every 1B tile changes a 1A target.
@@ -12,6 +14,7 @@ desitarget Change Log
     * We don't write any tile information when `ext=True` anyway.
 
 .. _`PR #879`: https://github.com/desihub/desitarget/pull/879
+.. _`PR #880`: https://github.com/desihub/desitarget/pull/880
 
 4.7.1 (2026-04-23)
 ------------------